Create gomod import annotation use --dependency instead of --src argument in generator fix enum clashes make more robust package name
