+++ .gitignore
@@ -1,3 +1,4 @@
 venv*
 .idea
 _examples/AddressBook/Examples
+.vscode
