+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # Changelog #
 All notable changes to this project will be documented in this file.
+
