+++ generator/templates/Enum.go.template
@@ -1,12 +1,12 @@
 // Code generated by goqface. DO NOT EDIT.
-package {{module.module.name_parts[-1]|lower}}
+package {{''.join(module.module.name_parts)|lower}}
 
 {% for enum in module.enums: %}
 type {{enum.name}} int
 
 const (
 {% for member in enum.members %}
-{{member.name}} = {{member.value}}
+{{member.unique_name}} = {{member.value}}
 {% endfor %}
 )
 {% endfor %}
