+++ generator/templates/Interface.go.template
@@ -1,5 +1,5 @@
 // Code generated by goqface. DO NOT EDIT.
-package {{module.module.name_parts[-1]|lower}}
+package {{''.join(module.module.name_parts)|lower}}
 import (
     "github.com/godbus/dbus/v5"
 {% for key, value in module.interface_imports.items() %}
