+++ generator/templates/Struct.go.template
@@ -1,5 +1,5 @@
 // Code generated by goqface. DO NOT EDIT.
-package {{module.module.name_parts[-1]|lower}}
+package {{''.join(module.module.name_parts)|lower}}
 
 import (
 {% for key, value in module.struct_imports.items() %}
