+++ tests/Phone/dependency/AddressBook.qface
@@ -1,4 +1,3 @@
-@gomod: "github.com/idleroamer/goqface/tests/Phone/dependency/Tests/Dependency/AddressBook"
 module Tests.Dependency.AddressBook 1.0;
