+++ tests/Phone/dependency/generator.go
@@ -1,4 +1,4 @@
 package test
 
-//go:generate python3 ../../../codegen.py --input AddressBook.qface
+//go:generate python3 ../../../generator/codegen.py --input AddressBook.qface
 //go:generate gofmt -w Tests
