+++ tests/Phone/dependent/dependent_test.go
@@ -2,7 +2,7 @@ package test
 
 import phone "github.com/idleroamer/goqface/tests/Phone/dependent/Tests/Phone"
 
-//go:generate python3 ../../../codegen.py --src ../ --input dependent/Phone.qface
+//go:generate python3 ../../../generator/codegen.py --dependency ../dependency --input Phone.qface
 //go:generate gofmt -w Tests
 
 type PhoneImpl struct {
